feat: split config.sh

+++ pypi-mirror.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 
-KUBESPRAY_DIR=${KUBESPRAY_DIR:-./kubespray}
+source config.sh
+
+if [ ! -e $KUBESPRAY_DIR ]; then
+    echo "No kubespray dir at $KUBESPRAY_DIR"
+    exit 1
+fi
+
+source ./.venv/bin/activate
 
 echo "==> Create pypi mirror for kubespray"
 pypi-mirror download -d outputs/pypi/files -r ${KUBESPRAY_DIR}/requirements.txt
